Add Online Wallet ZaloPaymet

+++ GenZStyleAPP.BAL/DTOs/Users/RegisterRequest.cs
@@ -19,7 +19,7 @@ namespace GenZStyleAPP.BAL.DTOs.Users
         public string? LastName { get; set; }
         public string? City { get; set; }
 
-        public int? Height { get; set; }
+        public decimal? Height { get; set; }
         public string? FullName { get; set; }
         public string? Address { get; set; }
         public string Phone { get; set; }
